requirements.txt
langchain
openai
claude
# gemini
langchain-openai
langchain-anthropic
# langchain-gemini
python-dotenv
langchain-community
sqlalchemy
psycopg2-binary
grandalf